PIL Resize
Implemented Resize using the PIL built in resize function. Also added FONT_RATIO and SCALE_FACTOR constants. SCALE_FACTOR is for resizing, which determines how much the resize should be, and FONT_RATIO is to account for the fact that font characters do not have the same width and height. This depends on font.